pyproject.toml
[project]
name = "karma"
version = "0.1.0"
description = "Karma — observability glue connecting Logfire and Langfuse via karma_code"
requires-python = ">=3.11"
dependencies = [
    "logfire>=4.25.0",
    "langfuse>=3.14.5",
    "python-dotenv>=1.2.2",
    "mcp>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
